reorganised for allowing more prompts
