+++ plugins/woocommerce-admin/client/customize-store/design-with-ai/prompts/index.ts
@@ -1,0 +1,2 @@
+export * from './colorChoices';
+export * from './lookAndTone';
